add tests and use size_t for portability

+++ levenshtein.h
@@ -5,7 +5,7 @@
  * MIT licensed.
  * Copyright (c) 2015 Titus Wormer <tituswormer@gmail.com> */
 
-/* Returns an unsigned integer, depicting
+/* Returns a size_t, depicting
  * the difference between `a` and `b`.
  * See http://en.wikipedia.org/wiki/Levenshtein_distance
  * for more information.
@@ -14,11 +14,11 @@
 extern "C" {
 #endif
 
-unsigned int
+size_t
 levenshtein(const char *a, const char *b);
 
-unsigned int
-levenshtein_n (const char *a, const unsigned int length, const char *b, const unsigned int bLength);
+size_t
+levenshtein_n (const char *a, const size_t length, const char *b, const size_t bLength);
 
 #ifdef __cplusplus
 }
